style: split util.h into smaller files
