+++ src/game/effects/dino_stomp.c
@@ -1,7 +1,6 @@
 #include "game/effects/dino_stomp.h"
 
 #include "global/vars.h"
-#include "util.h"
 
 #define MAX_BOUNCE 100
