+++ src/game/effects/waterfall.c
@@ -4,7 +4,6 @@
 #include "game/game.h"
 #include "game/items.h"
 #include "global/vars.h"
-#include "util.h"
 
 void SetupWaterfall(OBJECT_INFO *obj)
 {
