+++ src/game/laramisc.c
@@ -13,6 +13,7 @@
 #include "global/const.h"
 #include "global/types.h"
 #include "global/vars.h"
+#include "log.h"
 #include "specific/sndpc.h"
 
 #include <stddef.h>
