+++ src/game/setup.c
@@ -79,6 +79,7 @@
 #include "game/traps/thors_hammer.h"
 #include "global/const.h"
 #include "global/vars.h"
+#include "log.h"
 #include "specific/file.h"
 #include "specific/init.h"
 #include "specific/output.h"
