+++ src/specific/file.c
@@ -6,12 +6,13 @@
 #include "game/items.h"
 #include "game/setup.h"
 #include "global/vars.h"
+#include "log.h"
 #include "specific/hwr.h"
 #include "specific/init.h"
 #include "specific/smain.h"
 #include "specific/sndpc.h"
-#include "util.h"
 
+#include <stdio.h>
 #include <windows.h>
 
 int32_t MeshCount = 0;
