+++ src/specific/frontend.c
@@ -6,6 +6,7 @@
 #include "global/types.h"
 #include "global/vars.h"
 #include "global/vars_platform.h"
+#include "log.h"
 #include "specific/clock.h"
 #include "specific/display.h"
 #include "specific/file.h"
