+++ src/specific/shell.c
@@ -12,6 +12,7 @@
 #include "global/const.h"
 #include "global/types.h"
 #include "global/vars.h"
+#include "log.h"
 #include "specific/display.h"
 #include "specific/frontend.h"
 #include "specific/hwr.h"
