+++ src/util.h
@@ -1,9 +1,6 @@
 #ifndef T1M_UTIL_H
 #define T1M_UTIL_H
 
-#include <stdint.h>
-#include <stdio.h>
-
 #define SQUARE(A) ((A) * (A))
 #ifndef ABS
     #define ABS(x) (((x) < 0) ? (-(x)) : (x))
@@ -36,29 +33,5 @@
         (a) = (b);                                                             \
         (b) = (c);                                                             \
     } while (0)
-
-#pragma pack(push, 1)
-typedef struct {
-    uint8_t opcode; // must be 0xE9
-    uint32_t offset;
-} JMP;
-#pragma pack(pop)
-
-#define LOG_INFO(...) T1MLogFunc(__FILE__, __LINE__, __func__, __VA_ARGS__)
-#define LOG_ERROR(...) T1MLogFunc(__FILE__, __LINE__, __func__, __VA_ARGS__)
-#define LOG_DEBUG(...) T1MLogFunc(__FILE__, __LINE__, __func__, __VA_ARGS__)
-#define VAR_U_(address, type) (*(type *)(address))
-#define VAR_I_(address, type, value) (*(type *)(address))
-#define ARRAY_(address, type, length) (*(type(*) length)(address))
-
-void T1MLogFunc(
-    const char *file, int line, const char *func, const char *fmt, ...);
-void T1MInjectFunc(void (*from)(void), void (*to)(void));
-void T1MPrintStackTrace();
-
-#define INJECT(from, to)                                                       \
-    {                                                                          \
-        T1MInjectFunc((void (*)(void))from, (void (*)(void))to);               \
-    }
 
 #endif
